Fix freeing directory content blocks.
Previously, as we remove file from the directory, it's new content was
written back with the same length to overwrite last bytes of directory
contents with zeroes. In this case we can't check whether new directory
content occupies same amount of content blocks, or last content block
should be actually freed. We work-arounded this by writing directory
content twice: first time with old directory size to zero last bytes of
directory content, and second time with new (smaller) directory size to
check for possibly free last content block and clear it if that's the
case.

Signed-off-by: Kirill Leyfer <kolbasisha@gmail.com>